Update file name and move essence article repository to articlelist package etc...
