+++ app/src/main/java/com/example/androidstudyguide/articlelist/ArticleListFragment.kt
@@ -12,7 +12,6 @@ import androidx.lifecycle.ViewModelProvider
 import androidx.recyclerview.widget.DividerItemDecoration
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.example.androidstudyguide.data.remote.essence.EssenceAPI
-import com.example.androidstudyguide.data.remote.essence.EssenceArticleService
 import com.example.androidstudyguide.data.repository.ArticleRepository
 import com.example.androidstudyguide.databinding.FragmentArticleListBinding
 import com.example.androidstudyguide.models.Articles
@@ -26,7 +25,7 @@ class ArticleListFragment : Fragment(), ArticleClickListener {
     private val articleListViewModelFactory = object : ViewModelProvider.Factory {
         @Suppress("UNCHECKED_CAST")
         override fun <T : ViewModel> create(modelClass: Class<T>): T {
-            val articleRepository: ArticleRepository = EssenceArticleService(
+            val articleRepository: ArticleRepository = EssenceArticleRepository(
                 api = EssenceAPI.getInstance()
             )
             return ArticleListViewModel(articleRepository = articleRepository) as T
