+++ app/src/main/java/com/example/androidstudyguide/articlelist/EssenceArticleRepository.kt
@@ -1,10 +1,12 @@
-package com.example.androidstudyguide.data.remote.essence
+package com.example.androidstudyguide.articlelist
 
+import com.example.androidstudyguide.data.remote.essence.EssenceAPI
 import com.example.androidstudyguide.data.remote.essence.dto.EssenceFeedItemDto
 import com.example.androidstudyguide.data.repository.ArticleRepository
 import com.example.androidstudyguide.models.Articles
+import com.example.androidstudyguide.utils.EncodeString
 
-class EssenceArticleService(
+class EssenceArticleRepository(
     private val api: EssenceAPI
 ) : ArticleRepository {
     override suspend fun getArticles(): List<Articles> {
@@ -14,7 +16,7 @@ class EssenceArticleService(
 
 private fun EssenceFeedItemDto.toArticle(): Articles {
     return Articles(
-        title = this.title,
+        title = EncodeString(this.title),
         authorName = this.author.name,
         url = this.link.href
     )
